Relocated files and cleaned up connectionmanager
